pyproject.toml
[project]
name = "read2me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiofiles>=24.1.0",
  "cached-path>=1.6.6",
  "docling",
  "edge-tts>=7.0.0",
  "einops-exts>=0.0.4",
  "einops>=0.8.0",
  "fastapi>=0.115.6",
  "feedparser>=6.0.11",
  "feedsearch>=1.0.12",
  "langdetect>=1.0.9",
  "librosa>=0.10.2.post1",
  "llvmlite>=0.43.0",
  "munch>=4.0.0",
  "mutagen>=1.47.0",
  "newspaper4k>=0.9.3.1",
  "num2words>=0.5.14",
  "numpy",
  "ollama>=0.4.5",
  "openai>=1.59.3",
  "phonemizer>=3.3.0",
  "pillow",
  "playwright>=1.49.1",
  "pydub>=0.25.1",
  "pymupdf>=1.25.1",
  "python-dotenv>=1.0.1",
  "pytz>=2024.2",
  "soundfile>=0.13.0",
  "tldextract>=5.1.3",
  "torch",
  "torchaudio>=2.5.1",
  "trafilatura>=2.0.0",
  "transformers>=4.47.1",
  "tzlocal>=5.2",
  "uvicorn>=0.34.0",
  "werkzeug==2.0.3",
  "wikipedia>=1.4.0",
  "txtsplit>=1.0.0",
  "fastapi-cors>=0.0.6",
  "atproto>=0.0.57",
]
